refactored email template and theme API/ops - removed filesystem operations from lib
