+++ src/test/mocks/snapshotResolve.ts
@@ -1,6 +1,3 @@
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const path = require('path');
-
 module.exports = {
   resolveSnapshotPath: (testPath, snapshotExtension) => {
     let snapshotFilePath = '';
